More fixes for settings and disable by default
It doesn't seem to be necessary to specify options in initializationOptions
as the server will also request them per-file. And the server really
expects the same options (bar the "deno." prefix) in both cases so it
wasn't a complete solution anyway.

Also set "enabled": false by default with the intention of user enabling
it per-project.

And fixup some settings:
 - allow "null" value for some that had that value
 - rephrase "Not recommended to be set globally" since it has a bit
   different meaning in our case
 - remove "deno.path" which was an VS extension setting and in our case
   the "command" handles that